RUBY-1562 Use one file per class in crud spec runner
